Refactor #98: Consolidate auth options into cmd/ucm/utils/auth_options.go
Mirror cmd/root/auth_options.go: keep the SkipPrompt/SkipLoadUcm
setter+reader pairs (and the small profileFlagValue helper) together
in one file rather than split across auth.go (setters) and
configure.go (readers). No behavior change.
